Automatically increase `step` for large searches
If you have a very large report searches (for example 20000 items) then
it should only send a capped amount of AJAX requests to prevent it from
spamming the database (eventually leading to Internal Server Errors).

Previously it would always send one request per 20 items (so for our
example it would send 1000 separate requests). This means that when you
have very large searches this will effectively DOS the server, and
eventually reach a point where it starts generating
"Error connecting to database: Can't create UNIX socket (24)".

This leaves that option (by setting `step` to a number) however by
default `step` will be set automatically to try and send 100 AJAX
requests. This is set to be generally enough requests that the loading
bar looks reasonable (occasionally you get unlucky and all the requests
get replied to together however that is just a flaw with this system of
progress reporting).

This doesn't fully resolve the issue with server errors, notably if you
do get a server error during the process (for example if the server is
busy with other stuff) it will still just sit on a nearly full loading
bar with no way of notifying the client. I have another PR planned for
that to retry on a failure, however this should make that PR much more
effective so I'm landing it first.