Move the `Ajax.Request` call to a new function
As well as making the `EPrints_Screen_Report_Loader->execute` function
clearer (by moving a pretty large block of code) this will also allow it
to be re-called recursively on failure.

This commit shouldn't affect functionality in any way, just moving the
existing contents into a new function, unfortunately there appear to be
too many tab changes for git to notice this.