Retry the AJAX request if it fails
Rather than just never finishing loading if a single AJAX request fails
(as the `count` will never match the `total_no`), it should instead
retry.

To try and prevent it from hammering an already struggling server this
makes it wait 100ms between any retry, there should potentially be a
backoff or retry cap to fully stop this if it repeatedly fails (such as
if the DB is offline).